# docs/swagger.yaml
basePath: /api
definitions:
...
  models.InviteRequest:
    properties:
      invitee:
        type: string
    type: object
  models.InviteResponse:
    properties:
      created_at:
        type: string
      invitee_username:
        type: string
    type: object
...
  models.SubMemberResponse:
    properties:
      joined_at:
        type: string
      username:
        type: string
    type: object
  models.SubRequest:
    properties:
      description:
        type: string
      name:
        type: string
      private:
        type: boolean
    type: object
  models.SubResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      owner:
        type: string
      private:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@cortex-api.com
    name: CodeAndCraft Online
    url: https://github.com/CodeAndCraft-Online/cortex-api
  description: Cortex API is a Reddit-like social media platform backend API built
    with Go and PostgreSQL.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Cortex API
  version: "1.0"
paths:
...
  /sub/{subID}:
    delete:
      description: Deletes a subreddit completely (only sub owner can delete, cascade
        deletes memberships, posts, comments)
      parameters:
      - description: Sub ID
        in: path
        name: subID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Sub deleted successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Only sub owner can delete'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Sub not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Internal server error'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a sub
      tags:
      - Subs
    patch:
      consumes:
      - application/json
      description: Updates a subreddit's details (only sub owner can update description
        and privacy)
      parameters:
      - description: Sub ID
        in: path
        name: subID
        required: true
        type: string
      - description: Updated sub data (only description and private fields)
        in: body
        name: sub
        required: true
        schema:
          $ref: '#/definitions/models.SubRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated sub details
          schema:
            $ref: '#/definitions/models.SubResponse'
        "400":
          description: 'error: Bad request or invalid data'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Only sub owner can update'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Sub not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Internal server error'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a sub
      tags:
      - Subs
  /sub/{subID}/members:
    get:
      description: 'Retrieves all members of a subreddit (public subs: anyone, private
        subs: members/owners only)'
      parameters:
      - description: Sub ID
        in: path
        name: subID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Array of sub members with usernames and join dates
          schema:
            items:
              $ref: '#/definitions/models.SubMemberResponse'
            type: array
        "401":
          description: 'error: Unauthorized access to private sub'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Sub not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Internal server error'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get sub members
      tags:
      - Subs
  /sub/{subID}/pending-invites:
    get:
      description: Retrieves all pending invitations for a subreddit (only sub owners
        can view)
      parameters:
      - description: Sub ID
        in: path
        name: subID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Array of pending invites with usernames and creation dates
          schema:
            items:
              $ref: '#/definitions/models.InviteResponse'
            type: array
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Only sub owner can view invites'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Sub not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Internal server error'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get pending invites
      tags:
      - Subs
  /sub/sub/{subID}/invite:
    post:
      consumes:
      - application/json
      description: Allows sub owners and moderators to invite users to private subs
      parameters:
      - description: Sub ID
        in: path
        name: subID
        required: true
        type: string
      - description: User to invite
        in: body
        name: invite
        required: true
        schema:
          $ref: '#/definitions/models.InviteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Invitation sent'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Sub not found or permission denied'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Invite user to private sub
      tags:
      - Subs
  /sub/sub/{subID}/join:
    post:
      description: Allows users to join a subreddit (public or invited private subs)
      parameters:
      - description: Sub ID
        in: path
        name: subID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'membership.SubID: ID of joined sub'
          schema: {}
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Internal server error'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Join a sub
      tags:
      - Subs
  /sub/sub/{subID}/leave:
    post:
      description: Allows users to leave a subreddit they are currently a member of
      parameters:
      - description: Sub ID
        in: path
        name: subID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Left [sub name]'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Sub not found or not a member'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Internal server error'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Leave a sub
      tags:
      - Subs
  /sub/sub/{subID}/postCount:
    get:
      description: Returns the total number of posts in a specified sub
      parameters:
      - description: Sub ID
        in: query
        name: subID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Post count in the sub
          schema:
            type: integer
        "401":
          description: 'error: Unauthorized access to private sub'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Sub not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Internal server error'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get post count for sub
      tags:
      - Subs
  /sub/sub/{subID}/posts:
    get:
      description: Fetches all posts for a specific subreddit
      parameters:
      - description: Sub ID
        in: path
        name: subID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Array of posts in the sub
          schema:
            items: {}
            type: array
        "401":
          description: 'error: Unauthorized access to private sub'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Internal server error'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get posts by sub ID
      tags:
      - Subs
  /subs/:
    get:
      description: Returns all public subs and private subs the user is authorized
        to access
      produces:
      - application/json
      responses:
        "200":
          description: Array of available subs
          schema:
            items: {}
            type: array
        "500":
          description: 'error: Internal server error'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all subs
      tags:
      - Subs
    post:
      consumes:
      - application/json
      description: Creates a new subreddit (community)
      parameters:
      - description: Sub creation data
        in: body
        name: sub
        required: true
        schema:
          $ref: '#/definitions/models.SubRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created sub with ID and details
          schema: {}
        "400":
          description: 'error: Bad request or sub name already taken'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Internal server error'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new sub
      tags:
      - Subs
  /user/invite/{inviteID}/accept:
    post:
      description: Allows users to accept invitations to join private subs
      parameters:
      - description: Invite ID
        in: path
        name: inviteID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: You have joined the sub'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Invitation not found or expired'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Internal server error'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Accept sub invitation
      tags:
      - Subs
...
securityDefinitions:
  BearerAuth:
    description: '"JWT Authorization header using the Bearer scheme. Example: \"Authorization:
      Bearer {token}\""'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"